+++ backend/gonzales/static/assets/trending-down-DgdVng-S.js
@@ -1,4 +1,4 @@
-import{c as n}from"./index-Dtb8KvYj.js";/**
+import{c as n}from"./index-Dax4jwwU.js";/**
  * @license lucide-react v0.563.0 - ISC
  *
  * This source code is licensed under the ISC license.